remove appendix
I'm actually not sure if this is right. Am I proving it semantically? I
still don't understand why we can't just repeat the same steps of the
proof within the system. Is there really a Godelian reason we can't do
this? Or is it more that when the set Delta is infinite, then the
approach doesn't work anymore? I'd like to figure these questions out,
and may post to Math SE soon. But it's not important for the exposition
(in fact, actively harmful since it's a distraction) and I don't have
much time, so I am cutting this for now.